Improve parsing/looping code
The code running on a string list (as opposed to the original file object)
wasn't ideal. using the for loop over a string list didn't really work because
the the lines aren't actually taken of a stack, so a number of in(ter)dependent
counter variables were necessary.

This commit uses the (somewhat less pythonic) approach of an indexed
list, which has the great advantage that the nested loops can increment
the same counter as the outer loop.