Change Scheme module include strategy (breaking!)
Add a layer scheme/oll-core and add the scheme directory
to the module load path. This makes it possible to load modules
with
  (use-modules (oll-core stack))
instead of
  (use-modules (oll-core scheme stack))

This makes for a generally much better input language,
but it must be clear that this will break any code that uses
one of the Scheme modules defined in oll-core.

It has been fixed for all references in the oll-core package itself,
but it will have to be updated in the other packages as well.
Additionally there has to be a release note for any *user* who
may have added them to input files.