use-preset and checked properties
With (use-preset) it can be asked within the function
whether the function should be applied or not according
to preset settings.
NOTE: It is up to the function to interpret and implement
the meaning of "applying" the function.

instance properties are now type-checked too.
Properties not present in the propset or failing the predicate
issue a warning but are used anyway.

For properties expecting a symbol? strings are implicitly
converted for nicer user interface.