Implement parent presets, better type checking
Presets can now recursively load values from parent presets.
When parents and (grand)children specify the same properties
children override/shadow parents.

There's now a function in place that implicitly
converts string property values to symbols if the
property is defined as symbol?
This makes it possible to pass symbol values directly
in a \with block, without the #' notation.
NOTE: This really works with symbol? only, not with
custom predicates that *also* require the value to
be a symbol.