Handle preset filtering
The function (use-preset) within any
with-property-set function gives information
if the function should be used based on the
presence of a 'preset property. (It is the
function's responsibility to make use of this
information or not.)

Filtering is possible based on
- require-preset
- use-only-presets (or without preset)
- ignore-presets

Filter lists can be set globally or for a
specific property set. This gives the power
for sophisticated set-ups, but may lead
to confusing results, which has to be documented
carefully.