Change options.ily to options.scm => _options.scm
Try to improve the encapsulation.
Scheme symbols that are defined in _options.scm which is
loaded from options.scm are only visible within options.scm.
Symbols defined-public in options.scm which is loaded from
init.ily are visible to the LilyPond parser in any later file.
However, for use in Scheme modules (the upcoming predicates.scm)
options.scm has to be used explicitly.
Hope that goes well ...)