Separate out alist-access.scm
The alist-access module now only handles the low-level
access to actual alists.
The new module named-alists will become the public interface
to accessing named alists (like alist-access before) but with
separation of concerns, dealing with the "named" part
(by not storing parser variables anymore but providing a
data structure in a closure).